pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = ["Spidz <85760686+spidzsaus@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
SQLAlchemy = "^2.0.27"
fastapi = "^0.110.0"
python-dotenv = "^1.0.1"
uvicorn = "^0.27.1"
asyncpg = "^0.29.0"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
